[mooc-course] Move course module to the mooc context
