+++ src/Backoffice/README.md
@@ -1,0 +1,1 @@
+Here we'll have our Backoffice modules :)
