+++ src/Mooc/Courses/Application/Create/CreateCourseCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Backoffice\Courses\Application\Create;
+namespace CodelyTv\Mooc\Courses\Application\Create;
 
 use CodelyTv\Shared\Domain\Bus\Command\Command;
 use CodelyTv\Shared\Domain\ValueObject\Uuid;
