+++ src/Mooc/Courses/Domain/CourseCreatedDomainEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Backoffice\Courses\Domain;
+namespace CodelyTv\Mooc\Courses\Domain;
 
 use CodelyTv\Shared\Domain\Bus\Event\DomainEvent;
