+++ src/Mooc/Courses/Domain/CourseDescription.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Backoffice\Courses\Domain;
+namespace CodelyTv\Mooc\Courses\Domain;
 
 use CodelyTv\Shared\Domain\ValueObject\StringValueObject;
