+++ src/Mooc/Courses/Domain/CourseRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Backoffice\Courses\Domain;
+namespace CodelyTv\Mooc\Courses\Domain;
 
 interface CourseRepository
 {
