+++ src/Mooc/Courses/Infraestructure/Persistence/CourseRepositoryMySql.php
@@ -2,10 +2,10 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Backoffice\Courses\Infraestructure\Persistence;
+namespace CodelyTv\Mooc\Courses\Infraestructure\Persistence;
 
-use CodelyTv\Backoffice\Courses\Domain\Course;
-use CodelyTv\Backoffice\Courses\Domain\CourseRepository;
+use CodelyTv\Mooc\Courses\Domain\Course;
+use CodelyTv\Mooc\Courses\Domain\CourseRepository;
 use CodelyTv\Shared\Infrastructure\Doctrine\DoctrineRepository;
 
 final class CourseRepositoryMySql extends DoctrineRepository implements CourseRepository
