+++ src/Mooc/Shared/Domain/Videos/VideoUrl.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Mooc\Shared\Domain;
+namespace CodelyTv\Mooc\Shared\Domain\Videos;
 
 use CodelyTv\Shared\Domain\ValueObject\StringValueObject;
