+++ src/Mooc/Steps/Domain/Challenge/ChallengeStep.php
@@ -4,13 +4,13 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Mooc\Steps\Domain\Challenge;
 
+use CodelyTv\Mooc\Shared\Domain\Lessons\LessonId;
 use CodelyTv\Mooc\Steps\Domain\Step;
 use CodelyTv\Mooc\Steps\Domain\StepEstimatedDuration;
 use CodelyTv\Mooc\Steps\Domain\StepId;
 use CodelyTv\Mooc\Steps\Domain\StepOrder;
 use CodelyTv\Mooc\Steps\Domain\StepPoints;
 use CodelyTv\Mooc\Steps\Domain\StepTitle;
-use CodelyTv\Shared\Domain\Lessons\LessonId;
 use DateTimeImmutable;
 
 final class ChallengeStep extends Step
