+++ src/Mooc/Steps/Infrastructure/Persistence/Video.VideoStep.orm.yml
@@ -4,7 +4,7 @@ CodelyTv\Mooc\Steps\Domain\Video\VideoStep:
 
   embedded:
     videoUrl:
-      class: CodelyTv\Mooc\Shared\Domain\VideoUrl
+      class: CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl
       columnPrefix: false
     text:
       class: CodelyTv\Mooc\Steps\Domain\Video\VideoStepText
