+++ src/Mooc/Videos/Application/Create/VideoCreator.php
@@ -4,14 +4,14 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Mooc\Videos\Application\Create;
 
-use CodelyTv\Mooc\Shared\Domain\VideoUrl;
+use CodelyTv\Mooc\Shared\Domain\Courses\CourseId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
 use CodelyTv\Mooc\Videos\Domain\Video;
 use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoRepository;
 use CodelyTv\Mooc\Videos\Domain\VideoTitle;
 use CodelyTv\Mooc\Videos\Domain\VideoType;
 use CodelyTv\Shared\Domain\Bus\Event\DomainEventPublisher;
-use CodelyTv\Shared\Domain\CourseId;
 
 final class VideoCreator
 {
