+++ src/Mooc/Videos/Infrastructure/Persistence/Video.orm.yml
@@ -21,5 +21,5 @@ CodelyTv\Mooc\Videos\Domain\Video:
       class: CodelyTv\Mooc\Videos\Domain\VideoTitle
       columnPrefix: false
     url:
-      class: CodelyTv\Mooc\Shared\Domain\VideoUrl
+      class: CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl
       columnPrefix: false
