+++ tests/src/Mooc/Shared/Domain/Courses/CourseIdMother.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Test\Backoffice\Course\Domain;
+namespace CodelyTv\Test\Mooc\Shared\Domain\Courses;
 
-use CodelyTv\Shared\Domain\CourseId;
+use CodelyTv\Mooc\Shared\Domain\Courses\CourseId;
 use CodelyTv\Test\Shared\Domain\UuidMother;
 
 final class CourseIdMother
