+++ tests/src/Mooc/Shared/Domain/Lessons/LessonIdMother.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Test\Mooc\Shared\Domain\Lessons;
 
-use CodelyTv\Shared\Domain\Lessons\LessonId;
+use CodelyTv\Mooc\Shared\Domain\Lessons\LessonId;
 use CodelyTv\Test\Shared\Domain\UuidMother;
 
 final class LessonIdMother
