+++ tests/src/Mooc/Shared/Domain/Videos/VideoUrlMother.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Test\Mooc\Videos\Domain;
+namespace CodelyTv\Test\Mooc\Shared\Domain\Videos;
 
-use CodelyTv\Mooc\Shared\Domain\VideoUrl;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
 use CodelyTv\Test\Shared\Domain\UrlMother;
 
 final class VideoUrlMother
