+++ tests/src/Mooc/Steps/Domain/Video/VideoStepMother.php
@@ -4,7 +4,8 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Test\Mooc\Steps\Domain\Video;
 
-use CodelyTv\Mooc\Shared\Domain\VideoUrl;
+use CodelyTv\Mooc\Shared\Domain\Lessons\LessonId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
 use CodelyTv\Mooc\Steps\Domain\StepEstimatedDuration;
 use CodelyTv\Mooc\Steps\Domain\StepId;
 use CodelyTv\Mooc\Steps\Domain\StepOrder;
@@ -12,13 +13,12 @@ use CodelyTv\Mooc\Steps\Domain\StepTitle;
 use CodelyTv\Mooc\Steps\Domain\Video\VideoStep;
 use CodelyTv\Mooc\Steps\Domain\Video\VideoStepQuestion;
 use CodelyTv\Mooc\Steps\Domain\Video\VideoStepText;
-use CodelyTv\Shared\Domain\Lessons\LessonId;
 use CodelyTv\Test\Mooc\Shared\Domain\Lessons\LessonIdMother;
+use CodelyTv\Test\Mooc\Shared\Domain\Videos\VideoUrlMother;
 use CodelyTv\Test\Mooc\Steps\Domain\StepEstimatedDurationMother;
 use CodelyTv\Test\Mooc\Steps\Domain\StepIdMother;
 use CodelyTv\Test\Mooc\Steps\Domain\StepOrderMother;
 use CodelyTv\Test\Mooc\Steps\Domain\StepTitleMother;
-use CodelyTv\Test\Mooc\Videos\Domain\VideoUrlMother;
 use CodelyTv\Test\Shared\Domain\DateTimeMother;
 use DateTimeImmutable;
