+++ tests/src/Mooc/Videos/Application/Find/VideoResponseMother.php
@@ -4,17 +4,17 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Test\Mooc\Videos\Application\Find;
 
-use CodelyTv\Mooc\Shared\Domain\VideoUrl;
+use CodelyTv\Mooc\Shared\Domain\Courses\CourseId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
 use CodelyTv\Mooc\Videos\Application\Find\VideoResponse;
 use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoTitle;
 use CodelyTv\Mooc\Videos\Domain\VideoType;
-use CodelyTv\Shared\Domain\CourseId;
-use CodelyTv\Test\Backoffice\Course\Domain\CourseIdMother;
+use CodelyTv\Test\Mooc\Shared\Domain\Courses\CourseIdMother;
+use CodelyTv\Test\Mooc\Shared\Domain\Videos\VideoUrlMother;
 use CodelyTv\Test\Mooc\Videos\Domain\VideoIdMother;
 use CodelyTv\Test\Mooc\Videos\Domain\VideoTitleMother;
 use CodelyTv\Test\Mooc\Videos\Domain\VideoTypeMother;
-use CodelyTv\Test\Mooc\Videos\Domain\VideoUrlMother;
 use CodelyTv\Test\Shared\Domain\DuplicatorMother;
 
 final class VideoResponseMother
