+++ tests/src/Mooc/Videos/Domain/VideoCreatedDomainEventMother.php
@@ -4,13 +4,14 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Test\Mooc\Videos\Domain;
 
-use CodelyTv\Mooc\Shared\Domain\VideoUrl;
+use CodelyTv\Mooc\Shared\Domain\Courses\CourseId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
 use CodelyTv\Mooc\Videos\Domain\VideoCreatedDomainEvent;
 use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoTitle;
 use CodelyTv\Mooc\Videos\Domain\VideoType;
-use CodelyTv\Shared\Domain\CourseId;
-use CodelyTv\Test\Backoffice\Course\Domain\CourseIdMother;
+use CodelyTv\Test\Mooc\Shared\Domain\Courses\CourseIdMother;
+use CodelyTv\Test\Mooc\Shared\Domain\Videos\VideoUrlMother;
 
 final class VideoCreatedDomainEventMother
 {
